Preparation changes for merging with upstream (#45)
This performs some code refactoring to make the upstream merge in the next step less complicated.

The refactoring is cosmetic in that it doesn't affect the generated test vectors.
It mainly splits out the initialization and byte serialization of the transaction (in TransactionBase) to differing functions for header, transparent and sapling parts of the transaction.
The Orchard parts are in the child classes (TransactionV5 and TransactionV6) and are not separated out into functions.